requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
attrs==25.3.0
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
fastapi==0.115.14
fastjsonschema==2.21.1
fonttools==4.55.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
narwhals==1.31.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
plotly==6.0.1
postgres==4.0
prompt_toolkit==3.0.48
psutil==6.0.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pure_eval==0.2.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyfiglet==1.0.2
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
pywin32==308
pyzmq==26.2.0
referencing==0.36.2
rpds-py==0.25.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2024.2
uvicorn==0.35.0
wcwidth==0.2.13
